BigDecimal display format (space as thousand separator), getting operations from network

+++ app/src/main/java/ru/sirius/siriuswallet/network/DataRepository.kt
@@ -57,7 +57,7 @@ class DataRepository {
                 it.category.type.typeLocalizedName,
                 R.drawable.ic_salary,
                 LocalDateTime.ofInstant(it.creationDate, ZoneId.of("UTC")),
-                it.amount
+                it.amount.toBigDecimal()
             )
         }
     }
